requirements.txt
torch>=2.2.0
torchvision>=0.17.0
tqdm>=4.66.0
PyYAML>=6.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
tensorboard>=2.14.0
pytest>=7.4.0
